trinity/experiments/deadline-tracker/requirements.txt
# No heavy dependencies; uses stdlib + python-dateutil
python-dateutil>=2.8
